\name{dat}
\alias{dat}
\docType{data}
\title{Data of illustrative example in paper (see below)}
\description{
Illustrative data example with 100 observations in two dimensions.
}
\usage{data(dat)}
\format{
  The format is:
 num [1:100, 1:2] 3.39 4.08 4.35 4.89 4.55 ...
}
\details{
Data are constructed to contain global as well as local outliers.
}
\source{
P. Filzmoser, A. Ruiz-Gazen, and C. Thomas-Agnan: Identification of local multivariate
outliers. Submitted for publication, 2012.
}
\references{
P. Filzmoser, A. Ruiz-Gazen, and C. Thomas-Agnan: Identification of local multivariate
outliers. Submitted for publication, 2012.
}
\examples{
data(dat)
plot(dat)
}
\keyword{datasets}
