\name{X}
\alias{X}
\docType{data}
\title{Data (X coordinate) of illustrative example in paper (see below)}
\description{
Illustrative data example with 100 values for the X coordinate.
}
\usage{data(X)}
\format{
  The format is:
 num [1:100] 3.72 5.1 3.33 2.13 4.42 ...
}
\details{
Data are constructed to contain global as well as local outliers.
}
\source{
P. Filzmoser, A. Ruiz-Gazen, and C. Thomas-Agnan: Identification of local multivariate
outliers. Submitted for publication, 2012.
}
\references{
P. Filzmoser, A. Ruiz-Gazen, and C. Thomas-Agnan: Identification of local multivariate
outliers. Submitted for publication, 2012.
}
\examples{
data(X)
data(Y)
plot(X,Y)
}
\keyword{datasets}
