\name{Y}
\alias{Y}
\docType{data}
\title{Data (Y coordinate) of illustrative example in paper (see below)}
\description{
Illustrative data example with 100 values for the Y coordinate.
}
\usage{data(Y)}
\format{
  The format is:
 num [1:100] 1.25 1.4 0.372 0.791 2.74 ...
}
\details{
Data are constructed to contain global as well as local outliers.
}
\source{
P. Filzmoser, A. Ruiz-Gazen, and C. Thomas-Agnan: Identification of local multivariate
outliers. Submitted for publication, 2012.
}
\references{
P. Filzmoser, A. Ruiz-Gazen, and C. Thomas-Agnan: Identification of local multivariate
outliers. Submitted for publication, 2012.
}
\examples{
data(X)
data(Y)
plot(X,Y)
}
\keyword{datasets}
